Modify displayBooks() to focus on new entries
Initial iteration of the displayBooks() function looped through the
myLibrary array to create a new table containing every book in the
array. This isn't a problem if you only needed to run displayBooks once.
However, we want the function to run everytime a new book is added to
the library array. To solve this:
- Removed the table creation section. Instead, a blank table has been
added to index.html.
- The function now only looks at the last element of the library array.
- Added a data attribute to each row for future use when implementing
additional functionality.