Create input form dialog and apply page styles
Created a dialog that displays input fields to add a book to the library
table. Currently no client-side validation is applied and the form
doesn't get sent anywhere yet. This functionality to be added in a
subsequent update.

Applied a basic page styling via CSS. Page color palette not finalised.