Update theme, form style, and add form validation
Added some simple form validation for the text and number inputs to
ensure the inputs aren't blank and that the number of pages can't be
fewer than 0.

Updated the theme colors for the page to make everything more consistent
and readable. Additionally, finalised the layout for the dialog form and
added consistent styling to the form elements.
The only exception to this was the checkbox. I tried adding a custom
checkbox to the form in two different way; the first by adding an empty
span element and the second by adding content after the label. However,
despite managing to get these to work in a separate document, I couldn't
get them to function within the form set up in the dialog box. I suspect
some issues were being caused by the dialog box being in an absolute
position, but would need to investigate further. Dropped this feature
for now.