Refactor object contructors to classes
The following changes have been implemented:
- Refactored the Book object constructor into an object class.
- Moved the addBookToLibrary, toggleRead, and displayBook functions into
the class as static methods.
- The underlying library array (myLibrary) has also been moved into the
Book class. Originally, myLibrary had been declared as a global variable
but by putting it as a private static array inside the Book class, it
safeguards it from external manipulation.
- Moved the entire script into a factory function. This factory function
is immediately called and returns nothing. This ensures that the UI is
the only thing that can manipulate the library array.